% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/znajdz_tytul_jezyk.R
\name{znajdz_tytul_jezyk}
\alias{znajdz_tytul_jezyk}
\title{Znajdowanie linku do napisow filmowych ze strony opensubtitlers.org}
\usage{
znajdz_tytul_jezyk(tytul, jezyk)
}
\arguments{
\item{tytul}{zmienna typu character, case-insensitive, wskazujaca napisy jakiego filmu nas interesuja.}

\item{jezyk}{szukanych napisow.}
}
\value{
link do strony z napisami
}
\description{
Funkcja \code{znajdz_tytul_jezyk} jest rozszerzonym pomyslem kolezanki Ady :) Jej zadaniem jest generowanie linku do
strony z napisami wskazanego filmu we wskazanym jezyku.
}
\details{
Mozliwe parametry zmiennej \code{jezyk}:
\code{all} - wszystkie |
\code{pol} - polski |
\code{afr} - afrikaans |
\code{alb} - albanski |
\code{eng}  - angielski |
\code{ara} - arabski |
\code{arm} - armenski |
\code{baq} - basque |
\code{bel} - bialoruski |
\code{ben} - bengali |
\code{bos} - bosniacki |
\code{bre} - breton |
\code{bul} - bulgarski |
\code{bur} - burmese |
\code{chi} - chinski uproszczony |
\code{hrv} - chorwacki |
\code{cze} - czeski |
\code{dan} - dunski |
\code{dut} - niemiecki |
\code{epo} - esperanto |
\code{est} - estonski |
\code{fin} - finski |
\code{fre} - francuski |
\code{glg} - galicyjski |
\code{geo} - georgian |
\code{ell} - grecki |
\code{heb} - hebrajski |
\code{hin} - hinduski |
\code{spa} - hiszpanski |
\code{ice} - icelandic |
\code{ind} - indonezyjski |
\code{jpn} - japonski |
\code{cat} - katalonski |
\code{kaz} - kazachstanski |
\code{khm} - khmer |
\code{kor} - koreanski |
\code{lit} - litewski |
\code{ltz} - luksemburski |
\code{lav} - lotyski |
\code{mac} - macedonski |
\code{may} - malay |
\code{mal} - malayalam |
\code{mni} - manipuri |
\code{mon} - mongolian |
\code{mne} - montenegrin |
\code{ger} - niemiecki |
\code{nor} - norweski |
\code{oci} - occitan |
\code{per} - perski |
\code{por} - portugalski |
\code{rum} - rumunski |
\code{rus} - rosyjski |
\code{scc} - serbski |
\code{sin} - sinhalese |
\code{slo} - slowacki |
\code{slv} - slowenski |
\code{swa} - swahili |
\code{syr} - syriac |
\code{swe} - szwedzki |
\code{tgl} - tagalog |
\code{tha} - tajlandzki |
\code{tam} - tamil |
\code{tel} - telugu |
\code{tur} - turecki |
\code{ukr} - ukrainski |
\code{urd} - urdu |
\code{hun} - wegierski |
\code{vie} - wietnamski |
\code{ita} - wloski  |
}
\examples{
znajdz('titanic')
znajdz_tytul_jezyk('titanic','pol')
}
\author{
wersja 1.0.0: Adrianna Sudol, wersja 1.1.0: Katarzyna Fak
}

